requirements.txt
ImageHash==3.4
numpy==1.11.1
Pillow==3.3.1
